// src/Microsoft/Xna/Framework/Input/Enums.ts
export enum ButtonState {
  Released = 0,
  Pressed = 1,
}

/** XNA's physical and virtual game-pad button bit mask. */
export enum Buttons {
  DPadUp = 0x00000001,
  DPadDown = 0x00000002,
  DPadLeft = 0x00000004,
  DPadRight = 0x00000008,
  Start = 0x00000010,
  Back = 0x00000020,
  LeftStick = 0x00000040,
  RightStick = 0x00000080,
  LeftShoulder = 0x00000100,
  RightShoulder = 0x00000200,
  BigButton = 0x00000800,
  A = 0x00001000,
  B = 0x00002000,
  X = 0x00004000,
  Y = 0x00008000,
  LeftThumbstickLeft = 0x00200000,
  RightTrigger = 0x00400000,
  LeftTrigger = 0x00800000,
  RightThumbstickUp = 0x01000000,
  RightThumbstickDown = 0x02000000,
  RightThumbstickRight = 0x04000000,
  RightThumbstickLeft = 0x08000000,
  LeftThumbstickUp = 0x10000000,
  LeftThumbstickDown = 0x20000000,
  LeftThumbstickRight = 0x40000000,
}

export enum GamePadDeadZone {
  None = 0,
  IndependentAxes = 1,
  Circular = 2,
}

export enum GamePadType {
  Unknown = 0,
  GamePad = 1,
  Wheel = 2,
  ArcadeStick = 3,
  FlightStick = 4,
  DancePad = 5,
  Guitar = 6,
  AlternateGuitar = 7,
  DrumKit = 8,
  BigButtonPad = 0x300,
}

export enum KeyState {
  Up = 0,
  Down = 1,
}

/** XNA 4.0 keyboard values, which follow Windows virtual-key codes. */
export enum Keys {
  None = 0,
  Back = 8,
  Tab = 9,
  Enter = 13,
  Pause = 19,
  CapsLock = 20,
  Kana = 21,
  Kanji = 25,
  Escape = 27,
  ImeConvert = 28,
  ImeNoConvert = 29,
  Space = 32,
  PageUp = 33,
  PageDown = 34,
  End = 35,
  Home = 36,
  Left = 37,
  Up = 38,
  Right = 39,
  Down = 40,
  Select = 41,
  Print = 42,
  Execute = 43,
  PrintScreen = 44,
  Insert = 45,
  Delete = 46,
  Help = 47,
  D0 = 48,
  D1 = 49,
  D2 = 50,
  D3 = 51,
  D4 = 52,
  D5 = 53,
  D6 = 54,
  D7 = 55,
  D8 = 56,
  D9 = 57,
  A = 65,
  B = 66,
  C = 67,
  D = 68,
  E = 69,
  F = 70,
  G = 71,
  H = 72,
  I = 73,
  J = 74,
  K = 75,
  L = 76,
  M = 77,
  N = 78,
  O = 79,
  P = 80,
  Q = 81,
  R = 82,
  S = 83,
  T = 84,
  U = 85,
  V = 86,
  W = 87,
  X = 88,
  Y = 89,
  Z = 90,
  LeftWindows = 91,
  RightWindows = 92,
  Apps = 93,
  Sleep = 95,
  NumPad0 = 96,
  NumPad1 = 97,
  NumPad2 = 98,
  NumPad3 = 99,
  NumPad4 = 100,
  NumPad5 = 101,
  NumPad6 = 102,
  NumPad7 = 103,
  NumPad8 = 104,
  NumPad9 = 105,
  Multiply = 106,
  Add = 107,
  Separator = 108,
  Subtract = 109,
  Decimal = 110,
  Divide = 111,
  F1 = 112,
  F2 = 113,
  F3 = 114,
  F4 = 115,
  F5 = 116,
  F6 = 117,
  F7 = 118,
  F8 = 119,
  F9 = 120,
  F10 = 121,
  F11 = 122,
  F12 = 123,
  F13 = 124,
  F14 = 125,
  F15 = 126,
  F16 = 127,
  F17 = 128,
  F18 = 129,
  F19 = 130,
  F20 = 131,
  F21 = 132,
  F22 = 133,
  F23 = 134,
  F24 = 135,
  NumLock = 144,
  Scroll = 145,
  LeftShift = 160,
  RightShift = 161,
  LeftControl = 162,
  RightControl = 163,
  LeftAlt = 164,
  RightAlt = 165,
  BrowserBack = 166,
  BrowserForward = 167,
  BrowserRefresh = 168,
  BrowserStop = 169,
  BrowserSearch = 170,
  BrowserFavorites = 171,
  BrowserHome = 172,
  VolumeMute = 173,
  VolumeDown = 174,
  VolumeUp = 175,
  MediaNextTrack = 176,
  MediaPreviousTrack = 177,
  MediaStop = 178,
  MediaPlayPause = 179,
  LaunchMail = 180,
  SelectMedia = 181,
  LaunchApplication1 = 182,
  LaunchApplication2 = 183,
  OemSemicolon = 186,
  OemPlus = 187,
  OemComma = 188,
  OemMinus = 189,
  OemPeriod = 190,
  OemQuestion = 191,
  OemTilde = 192,
  ChatPadGreen = 202,
  ChatPadOrange = 203,
  OemOpenBrackets = 219,
  OemPipe = 220,
  OemCloseBrackets = 221,
  OemQuotes = 222,
  Oem8 = 223,
  OemBackslash = 226,
  ProcessKey = 229,
  OemCopy = 242,
  OemAuto = 243,
  OemEnlW = 244,
  Attn = 246,
  Crsel = 247,
  Exsel = 248,
  EraseEof = 249,
  Play = 250,
  Zoom = 251,
  Pa1 = 253,
  OemClear = 254,
}